requirements.txt
numpy
pandas
pyserial